Add a setting to force LPAR to set defaults in NVRAM before booting
This commit introduces the `SET_NVRAM_DEFAULTS` setting, that will cause
the functions in lib/bootloader_pvm.pm to boot the LPAR into the Open
Firmware command line prompt and issue a `SET_NVRAM_DEFAULTS` to reset
the values in the LPAR's NVRAM to its defaults. This can help to fix
some boot issues caused by aborted boot processes in the LPAR, which
in turn prevent the GRUB menu from being loaded. If set, LPAR will boot
into the Open Firmware prompt, issue the command and then continue to
boot normally. If not set, it will boot as usual.