Changes to enable security_pam tests on SLE>=16
A couple of tests are skipped. These are:

- pam_login
- pam_mount

The pam_login would rely on the pam_ssh module which was removed from SLE 16.

The pam_mount test would need the pam_mount module which is not included in the distro.
The package is not even on the wishlist of SLE 16.

Other tests have been modified -where needed- to run on SLE>=16.

Signed-off-by: Ferenc Szekely <feri.szekely@suse.com>