Add workaround allowing to install certs on SLE12-SP5
Given that our primary need was simply to install certificates, we decided to abandon
he complex logic that determined which package version to select for each run.
Our new approach is to install the TW package universally. Regrettably,
this has presented a challenge with SLE 12 SP5, as the TW package utilizes an unsupported compression method.